refactor: refactor tools into folders
