--- packages/blade-mcp/src/tools/figma-to-code/index.ts
@@ -1,1 +1,0 @@
-export { registerFigmaToCodeTool } from './figma-to-code.js';
