+++ packages/blade-mcp/src/utils/constants.ts
@@ -1,6 +1,8 @@
-export const CONFIG = {
+const URLS = {
   FIGMA_TO_CODE_URL: {
     DEV: 'http://localhost:8888',
     PROD: 'https://blade-chat-base.dev.razorpay.in',
   },
 } as const;
+
+export { URLS };
